feat: Refactor dropdown menu to be more reusable
* refactored settings menu

* finished dropdown refactor

* fix: settings menu height

---------
