+++ .cargo/config.toml
@@ -1,3 +1,3 @@
-[target.x86_64-unknown-linux-gnu]
-linker = "clang"
-rustflags = ["-C", "link-arg=-fuse-ld=lld"]
+# [target.'cfg(target_os="linux")']
+# linker = "clang"
+# rustflags = ["-C", "link-arg=-fuse-ld=lld"]
