+++ src/view/mod.rs
@@ -1,6 +1,6 @@
 pub mod bottom_bar;
 pub mod composite;
 pub mod docview;
-pub mod settings_menu;
+pub mod dropdown;
 pub mod sidebar;
 pub mod top_bar;
